Major refactor in source code file path to be in complain with `zeitwerk`
