+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2021-10-06 08:55:51 UTC using RuboCop version 1.22.1.
+# on 2021-10-08 11:09:45 UTC using RuboCop version 1.22.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -69,8 +69,8 @@ Rails/HasManyOrHasOneDependent:
 # ForbiddenMethods: decrement!, decrement_counter, increment!, increment_counter, insert, insert!, insert_all, insert_all!, toggle!, touch, touch_all, update_all, update_attribute, update_column, update_columns, update_counters, upsert, upsert_all
 Rails/SkipsModelValidations:
   Exclude:
+    - 'app/controllers/concerns/server_thread_methods/thread_manager.rb'
     - 'app/models/server.rb'
-    - 'app/server/managers/thread_manager.rb'
 
 # Offense count: 3
 # Configuration parameters: Include.
@@ -84,7 +84,7 @@ Rails/UniqueValidationWithoutIndex:
 # Offense count: 1
 Security/Open:
   Exclude:
-    - 'app/server/managers/html_result_manager.rb'
+    - 'app/controllers/concerns/server_thread_methods/html_result_manager.rb'
 
 # Offense count: 94
 # Configuration parameters: AllowedConstants.
